Managed to create a network graph
Finally managed to create a function in the Analysis.r file that creates a netword chart that visualize connection among Summons, Arrests, and police officers dispatched.

I also did minor cleaning in in the Parser.R file. I added a ID column that would easily allow a join back to the master data if we decide to create views.


Former-commit-id: 1cca3ffbab19dfe3a55282b3c296b11ec6b0e87b